requirements.txt
# Core dependencies
streamlit>=1.28.0
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-community>=0.0.20
chromadb>=0.4.22
openai>=1.12.0  # Opcionális, ha OpenAI-t is szeretnél használni

# Document processing
pypdf>=3.17.0
python-docx>=1.1.0
unstructured>=0.11.8

# Vector database and embeddings
sentence-transformers>=2.3.1
FlagEmbedding>=1.2.0  # BGE-M3 modellhez
faiss-cpu>=1.7.4  # Alternative to ChromaDB

# LLM - Qwen-4B lokális modell
transformers>=4.35.0
torch>=2.1.0
accelerate>=0.24.0
bitsandbytes>=0.41.0  # Opcionális, quantizációhoz

# Evaluation and metrics
numpy>=1.24.0
pandas>=2.1.0
scikit-learn>=1.3.0

# Monitoring and analytics
plotly>=5.18.0
matplotlib>=3.8.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.5.0
tqdm>=4.66.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0

# Optional: For advanced reranking
sentence-transformers>=2.3.1
torch>=2.1.0

